\name{SCEPtER-package}
\docType{package}
\alias{SCEPtER-package}
\title{A grid-based estimation tool for stellar ages, masses, and radii}
\description{
  The package estimates stellar mass, radius, and age given
  observational 
  data of effective temperature, [Fe/H], and astroseismic parameters.
  The results are obtained adopting a maximum likelihood
  technique on a grid of computed stellar models.
  
}
\references{
  G. Valle, M. Dell'Omodarme, P.G. Prada Moroni, and S. Degl'Innocenti
(2014). \emph{Uncertainties in grid-based estimates of stellar mass and
  radius. SCEPtER: Stellar CharactEristics Pisa
  Estimation gRid}. A&A, 561, A125.

G. Valle, M. Dell'Omodarme, P.G. Prada Moroni, and S. Degl'Innocenti
(2014). \emph{Uncertainties in asteroseismic grid-based estimates of stellar ages. SCEPtER: Stellar CharactEristics Pisa
  Estimation gRid}. A&A, accepted.
}
\keyword{package}
